# slices: use dual-pivot quicksort