… slices/zsortfunc.go
… slices/zsortordered.go