Dual-pivot quicksort has smaller recursion depth.
Some pattern-defeating technique is added to make this variant.

And use double reversion instead of blockswap for symmerge.

Benchmark results on Xeon-8372C
name                          old time/op   new time/op   delta
SlicesSortInts-4              7.01ms ± 0%  5.84ms ± 0%  -16.64%  (p=0.000 n=10+10)
SlicesSortInts_Sorted-4       80.1µs ± 1%  69.1µs ± 0%  -13.76%  (p=0.000 n=8+10)
SlicesSortInts_Reversed-4      116µs ± 1%   108µs ± 0%   -6.72%  (p=0.000 n=10+10)
SlicesSortInts_Mixed-4        5.36ms ± 0%  4.31ms ± 0%  -19.53%  (p=0.000 n=10+9)
SlicesSortStrings-4           22.5ms ± 0%  20.9ms ± 1%   -7.12%  (p=0.000 n=10+10)
SortFuncStructs-4             13.4ms ± 1%  12.5ms ± 1%   -7.00%  (p=0.000 n=10+10)
SortFuncStructs_Stable-4      31.1ms ± 1%  30.7ms ± 1%   -1.25%  (p=0.000 n=10+10)

For golang/go#52789